- Deleted Application test file that contained all the tests. - Created a folder inside tests for storing the integration tests, that directly test the app full functionality with a containerized PostgreSQL database. - Created a separate class for the related User integration tests. - Implemented all the User integration tests, with different situations to test that the database data always ends as expected. - Added a config test file, for testing general configurations of the application. Currently, it tests if the current PostgreSQL database is the same as the one defined with Exposed DSL. And if docker is available, wich is needed for the integration tests.
